Natural-Language/Conversational-Language-Understanding/requirements.txt
python-dotenv~=1.0.0
azure-identity>=1.15.0
python-dateutil
azure-ai-language-conversations